Refactor authentication handling: replace server-side AuthButton with client-side implementation and update layout to use SessionProvider
