+++ .gitignore
@@ -15,6 +15,7 @@
 
 # production
 /build
+package-lock.json
 
 # misc
 .DS_Store
@@ -27,6 +28,7 @@ yarn-error.log*
 
 # local env files
 .env*.local
+.env
 
 # vercel
 .vercel
